feat(core): extract WikiMemory services and test __testAccess
Split retrieval, import/export, embeddings, and write orchestration into
focused services behind a slim WikiMemory facade; rename internal job hooks
to doRun* / doImportEntity and add WikiMemory.__testAccess for typed tests.
