+++ packages/core/__tests__/write.test.ts
@@ -48,7 +48,7 @@ describe('write() — librarian trigger suppression', () => {
     await wiki.setup();
 
     (wiki as any).jobManager.activeMaintenanceJobs.add(`${PREFIX}:user-1:forget`);
-    const librarianSpy = vi.spyOn((wiki as any).maintenanceService, '_doRunLibrarian');
+    const librarianSpy = vi.spyOn(wiki.__testAccess.maintenanceService, 'doRunLibrarian');
 
     await wiki.write('user-1', { event_type: 'observation', summary: 'no librarian' });
 
@@ -64,7 +64,7 @@ describe('write() — librarian trigger suppression', () => {
     await wiki.setup();
 
     (wiki as any).jobManager.activeMaintenanceJobs.add(`${PREFIX}:user-1:prune`);
-    const librarianSpy = vi.spyOn((wiki as any).maintenanceService, '_doRunLibrarian');
+    const librarianSpy = vi.spyOn(wiki.__testAccess.maintenanceService, 'doRunLibrarian');
 
     await wiki.write('user-1', { event_type: 'observation', summary: 'no librarian prune' });
