+++ packages/core/src/index.ts
@@ -3,6 +3,7 @@ import type { SQLiteAdapter, WikiOptions } from './types';
 
 export * from './types';
 export { WikiMemory } from './WikiMemory';
+export type { WikiMemoryTestAccess } from './WikiMemory';
 export { formatContext } from './utils/formatContext';
 export { formatMemoryDump } from './utils/formatMemoryDump';
 export { parseEmbedding } from './utils/embedding';
